# functions/swagger.yaml
openapi: "3.0.0"
info:
  title: Tech App backend
  description: Documentation for the backend api
  version: 1.1.0
  contact:
    email: unknown39825@gmail.cmo
    url: "https://github.com/technobyte-nitkkr"
servers:
  - url : "http://localhost:5000/techspardha-87928/us-central1/api"
    description: localhost
  - url: "https://us-central1-techspardha-87928.cloudfunctions.net/api"
    description: server
...
components:
  securitySchemes:
    Authdetails:
      type: apiKey
      in: header
      name: Authorization
  schemas:
    simpleResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
    userInfo:
      type: object
      properties:
        email:
          type: string
        name:
          type: string
        picture:
          type: string
        onBoard:
          type: boolean
        college:
          type: string
        year:
          type: string
        admin:
          type: boolean
        role:
          type: string
    eventInfo:
      type: object
      properties:
        eventName:
          type: string
        eventCategory:
          type: string
        venue:
          type: string
        description:
          type: string
        endTime:
          type: number
        startTime:
          type: number
        file:
          type: string
        cordinators:
          type: array
          items:
            type: object
            properties: 
              coordinator_name:
                type: string
              coordinator_number:
                type: string
        rules:
          type: array
          items:
            type: string
paths:
...
  /user/event:
    get:
      tags:
        - User
      summary: get the events of loggedin user
      security:
        - Authdetails: []

      responses:
        200:
          description: All good success
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  data:
                    type: object
                    properties:
                      events:
                        type: array
                        items:
                          $ref: '#/components/schemas/eventInfo'
                          
        500:
          description: internal server error
    put:
      tags:
        - User
      summary: add event of loggedin user
      security:
        - Authdetails: []
      requestBody:
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              required:
                - eventName
                - eventCategory
              properties:
                eventName:
                  type: string
                eventCategory:
                  type: string
      responses:
        200:
          description: All good success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/simpleResponse'
        500:
          description: internal server error
...
  /events:
    get:
      tags:
        - Information
      summary: get the events names
      parameters:
        - in: query
          name: eventCategory
          schema:
            type: string
            enum:
              [
                Astronomy,
                Design,
                Informals,
                Managerial,
                Online-Events,
                Papyrus-Vitae,
                Programming,
                Quizzes,
                Robotics,
              ]
      responses:
        200:
          description: All good success
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  data:
                    type: object
                    properties:
                      events:
                        type: array
                        items:
                          type: object
                          properties:
                            eventName:
                              type: string
                            eventCategory:
                              type: string
        500:
          description: internal server error
    post:
      security:
        - Authdetails: []
      tags:
        - manager
      summary: add event for manager user
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required:
                - eventData
              properties:
                eventData:
                  $ref: '#/components/schemas/eventInfo'
      responses:
        200:
          description: All good success
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
        401:
          description: Not authenticated
        500:
          description: internal server error
...
  /timestamp/events:
    get:
      tags:
        - Information
      summary: get upcoming events
      parameters:
        - in: query
          name: timestamp
          schema:
            type: number
      responses:
        200:
          description: All good success
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  data:
                    type: object
                    properties:
                      events:
                        type: array
                        items: 
                          type: object
        500:
          description: internal server error
...
  /admin/event:
    get:
      tags:
        - manager
      summary: get the of all users
      security:
        - Authdetails: []
      parameters:
        - in: query
          name: eventCategory
          required: true
          schema:
            type: string
            enum:
              [
                Astronomy,
                Design,
                Informals,
                Managerial,
                Online-Events,
                Papyrus-Vitae,
                Programming,
                Quizzes,
                Robotics,
              ]
        - in: query
          name: eventName
          required: true
          schema:
            type: string
      responses:
        200:
          description: All good success
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  data:
                    type: object
                    properties:
                      users:
                        type: array
                        items: 
                          $ref: '#/components/schemas/userInfo'
        500:
          description: internal server error
...